requirements.txt
grpcio==1.54.0
protobuf==3.20.3
simple_term_menu==1.6.1
